challenge/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Luteach <equipo@luteach.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django-cors-headers = "^3.13.0"
Django = "^4.1.2"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
psycopg2 = "^2.9.3"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
